fix datatype

+++ src/gl/vbo.rs
@@ -48,7 +48,7 @@ pub fn vertex_attrib_pointer<T>(
 
 pub fn buffer_data<T>(n: u32, data: &[T], usage: u32) {
     let pointer = data.as_ptr() as *const std::ffi::c_void;
-    let size = (data.len() * std::mem::size_of::<T>()) as isize;
+    let size = (data.len() * std::mem::size_of::<T>()) as i64;
     unsafe {
         glBufferData(n, size, pointer, usage);
     }
